moved schemas to models
