+++ routes/api/contacts.js
@@ -4,27 +4,31 @@ const router = express.Router();
 
 const { ctrlContacts } = require("../../controllers");
 const { isValidId, validateBody } = require("../../middlewares");
-const { addContactSchema, updStatusContactSchema } = require("../../schemas");
+const { schemas } = require("../../models");
 
 router.get("/", ctrlContacts.getAll);
 
 router.get("/:contactId", isValidId, ctrlContacts.getContact);
 
-router.post("/", validateBody(addContactSchema), ctrlContacts.postContact);
+router.post(
+  "/",
+  validateBody(schemas.addContactSchema),
+  ctrlContacts.postContact
+);
 
 router.delete("/:contactId", isValidId, ctrlContacts.remContact);
 
 router.put(
   "/:contactId",
   isValidId,
-  validateBody(addContactSchema),
+  validateBody(schemas.addContactSchema),
   ctrlContacts.updContact
 );
 
 router.patch(
   "/:contactId/favorite",
   isValidId,
-  validateBody(updStatusContactSchema),
+  validateBody(schemas.updStatusContactSchema),
   ctrlContacts.updStatusContact
 );
