--- schemas/contacts.js
@@ -1,17 +1,0 @@
-const Joi = require("joi");
-
-const addContactSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-  favorite: Joi.boolean(),
-});
-
-const updStatusContactSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
-module.exports = {
-  addContactSchema,
-  updStatusContactSchema,
-};
