--- schemas/index.js
@@ -1,6 +1,0 @@
-const { addContactSchema, updStatusContactSchema } = require("./contacts.js");
-
-module.exports = {
-  addContactSchema,
-  updStatusContactSchema,
-};
